Partially solve day 9 part 1
Partially solve part 1 of day 9 for 2019.
VM updated to handle opcode 9 and relative mode, but doesn't give an output value for the Keycode other than zero, which is wrong.

+++ tests/AdventOfCode.Tests/Puzzles/Y2019/Day07Tests.cs
@@ -25,10 +25,10 @@ namespace MartinCostello.AdventOfCode.Puzzles.Y2019
         [InlineData("3,15,3,16,1002,16,10,16,1,16,15,15,4,15,99,0,0", 43210)]
         [InlineData("3,23,3,24,1002,24,10,24,1002,23,-1,23,101,5,23,23,1,24,23,23,4,23,99,0,0", 54321)]
         [InlineData("3,31,3,32,1002,32,10,32,1001,31,-2,31,1007,31,0,33,1002,33,7,33,1,33,31,31,1,32,31,31,4,31,99,0,0,0", 65210)]
-        public void Y2019_Day07_RunProgram_Returns_Correct_Output(string program, int expected)
+        public void Y2019_Day07_RunProgram_Returns_Correct_Output(string program, long expected)
         {
             // Act
-            int actual = Day07.RunProgram(program);
+            long actual = Day07.RunProgram(program);
 
             // Assert
             actual.ShouldBe(expected);
